rename db_service and import {prisma} from db_client
