+++ api/controllers/auth.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import prisma from "./db_service";
+import { prisma } from "./../utils/db_client";
 import * as jwt from "jsonwebtoken";
 import * as bcrypt from "bcryptjs";
 import { JWT_SECRET_KEY, expiresIn } from "../middlewares/checkJwt";
@@ -8,7 +8,42 @@ async function refreshToken(req: Request, res: Response) {
   // const { refresh_token } = req.query;
   //recieve an refresh token from and sends an access token
 
+  //if the acces token expired we need refresh token
+  // the refresh tocken will be sent along with the refresh token
+
+  //if (expiresIn<= now){
+  //send a new access token
+
   // res.json({ token: token, newToken });
+
+  // const token = req.cookies.jid;
+  //   if (!token) {
+  //     return res.send({ ok: false, accessToken: "" });
+  //   }
+
+  //   let payload: any = null;
+  //   try {
+  //     payload = verify(token, process.env.REFRESH_TOKEN_SECRET!);
+  //   } catch (err) {
+  //     console.log(err);
+  //     return res.send({ ok: false, accessToken: "" });
+  //   }
+
+  //   // token is valid and
+  //   // we can send back an access token
+  //   const user = await User.findOne({ id: payload.userId });
+
+  //   if (!user) {
+  //     return res.send({ ok: false, accessToken: "" });
+  //   }
+
+  //   if (user.tokenVersion !== payload.tokenVersion) {
+  //     return res.send({ ok: false, accessToken: "" });
+  //   }
+
+  //   sendRefreshToken(res, createRefreshToken(user));
+
+  //   return res.send({ ok: true, accessToken: createAccessToken(user) });
 
   console.log("refreshToken");
 }
@@ -25,27 +60,28 @@ async function login(req: Request, res: Response) {
     },
   });
   if (!userByEmail) {
-    res.status(401).json({ error: "Unotorised, User not found" });
+    return res.status(401).json({ error: "Unauthorised, User not found" });
   }
-  console.log("find user by email", userByEmail);
 
   // Check if password is correct
   try {
     const passwordsMatchCheck = await bcrypt.compare(
       req.body.password,
-      <string>userByEmail?.password
+      userByEmail.password!
     );
-    console.log("match", passwordsMatchCheck);
 
     if (passwordsMatchCheck) {
-      const accessToken = jwt.sign({ id: userByEmail?.id }, JWT_SECRET_KEY, {
+      const accessToken = jwt.sign({ id: userByEmail.id }, JWT_SECRET_KEY, {
         expiresIn,
       });
 
       // returns or gives the user a token(access token, refresh token)
-      res
-        .status(200)
-        .json({ user: userByEmail, succes: true, accessToken: accessToken });
+      res.status(200).json({
+        user: userByEmail,
+        succes: true,
+        accessToken: accessToken,
+        // , refreshToken: refreshToken //TODO diffrent secrets
+      });
     } else {
       res.status(401).json({ error: "Password incorrect" });
     }
@@ -57,8 +93,8 @@ async function login(req: Request, res: Response) {
 async function createUser(req: Request, res: Response) {
   console.log("createUser");
   try {
-    const { username, email } = req.body;
-    const password = await bcrypt.hash(req.body.password, 10);
+    const { username, email, password } = req.body;
+    const hashPassword = await bcrypt.hash(password, 10);
 
     // console.log("createUser", username, email, password);
     // email lookup
@@ -75,26 +111,31 @@ async function createUser(req: Request, res: Response) {
     });
 
     if (user) {
-      res.status(409).json({ error: "User already exists" });
-    } else {
-      user = await prisma.users.create({
-        data: {
-          username: username,
-          email: email,
-          password: password,
-        },
-      });
+      return res.status(409).json({ error: "User already exists" });
     }
 
-    const accessToken = jwt.sign(
-      { id: user.id, email: user.email },
-      JWT_SECRET_KEY,
-      {
-        expiresIn,
-      }
-    );
+    const newUser = await prisma.users.create({
+      data: {
+        username: username,
+        email: email,
+        password: hashPassword,
+      },
+      select: {
+        id: true,
+        email: true,
+      },
+    });
 
-    res.status(201).json({ user: user, access_token: accessToken, expiresIn });
+    //Only with login
+    // const accessToken = jwt.sign(
+    //   { id: user.id, email: user.email },
+    //   JWT_SECRET_KEY,
+    //   {
+    //     expiresIn,
+    //   }
+    // );
+
+    res.status(201).json({ user: newUser });
   } catch (db_error) {
     res.status(500).json(db_error);
   }
